refactor(web): decompose LocationContext provider (REF-505)
- Extract useLocationInit hook
- Extract useMultiTabLocationSync hook
- Reduce LocationContext.tsx from 506 to 357 LOC
- Preserve location status provider hierarchy and context hooks
