+++ apps/web/src/context/LocationContext.tsx
@@ -10,13 +10,10 @@ import {
     useCallback,
     useMemo,
 } from "react";
-import { getMe } from "@/lib/api/user/users";
-import { 
-    normalizeToAppLocation as normalizeLocation 
-} from "@/lib/location/locationService";
 import type { AppLocation, GeoJSONPoint } from "@/types/location";
 import { DEFAULT_APP_LOCATION } from "@/types/location";
 import { isGenericDetectedLocation } from "@/lib/location/locationService";
+
 // Hooks
 import {
     useLocationStorage,
@@ -25,6 +22,8 @@ import {
 } from "./hooks/useLocationStorage";
 import { useUnifiedLocationDetection } from "@/hooks/useUnifiedLocationDetection";
 import { useLocationActionHandlers } from "./hooks/useLocationActionHandlers";
+import { useLocationInit } from "./hooks/useLocationInit";
+import { useMultiTabLocationSync } from "./hooks/useMultiTabLocationSync";
 
 /* -------------------------------------------------------------------------- */
 /* TYPES */
@@ -69,8 +68,6 @@ export type LocationDispatchContextType = {
 
 export type LocationActionsContextType = LocationDispatchContextType;
 
-const LOCATION_PROMPT_DELAY_MS = 800;
-
 const getLocationStatus = (source: LocationData["source"]): LocationStatus =>
     source === "manual" ? "manual_selection" : "granted";
 
@@ -100,8 +97,6 @@ export function LocationProvider({
     const [_promptDismissed, setPromptDismissed] = useState(false);
     const [locationExpired, setLocationExpired] = useState(false);
 
-    const initializedRef = useRef(false);
-    const promptDelayTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
     const genericLocationRefreshKeyRef = useRef<string | null>(null);
     const autoDetectedRef = useRef(false);
 
@@ -120,7 +115,6 @@ export function LocationProvider({
         setPromptDismissed(dismissed);
         setPromptDismissedFlag(dismissed);
         
-        // Expiration logic is handled by setting a timestamp in localstorage instead of just a boolean
         if (dismissed && typeof window !== "undefined") {
             const expiry = Date.now() + 30 * 24 * 60 * 60 * 1000; // 30 days
             localStorage.setItem(LOCATION_PROMPT_DISMISSED_KEY + "_expiry", expiry.toString());
@@ -165,7 +159,7 @@ export function LocationProvider({
     }, [applyResolvedLocation, setPermissionBlockedFlag, logAnalytics]);
 
     const handleError = useCallback((msg: string) => {
-        setStatus(location.source === "manual" ? "manual_selection" : "prompt"); // Back to prompt so they can retry or manual
+        setStatus(location.source === "manual" ? "manual_selection" : "prompt");
         setDetectError(msg);
     }, [location.source]);
 
@@ -196,7 +190,6 @@ export function LocationProvider({
         }
         setStatus("checking");
         
-        // Fire analytics
         logAnalytics?.({
             source: 'default',
             city: 'Unknown',
@@ -208,7 +201,6 @@ export function LocationProvider({
         const result = await unifiedDetect({ persist, force });
 
         if (isAutoPrompt && result?.failure) {
-            // Check native browser state to see if it's STILL prompt (meaning user dismissed it instead of denying)
             if (typeof navigator !== "undefined" && navigator.permissions) {
                 try {
                     const permResult = await navigator.permissions.query({ name: 'geolocation' });
@@ -272,167 +264,27 @@ export function LocationProvider({
         void performReverseGeocode(lat, lng);
     }, [location, performReverseGeocode]);
 
-    // ── Initialization ────────────────────────────────────────────────────────
-    const hydrateProfileLocation = useCallback(async (): Promise<LocationData | null> => {
-        if (!initialHasAuthCookie) return null;
-        try {
-            const user = await getMe({ silent: true });
-            if (!user?.location?.city) return null;
+    // ── Initialization & Storage Sync Sub-Hooks ───────────────────────────────
+    useLocationInit({
+        initialHasAuthCookie,
+        readStoredLocation,
+        applyResolvedLocation,
+        readPermissionBlockedFlag,
+        setPermissionBlockedFlag,
+        setStatus,
+        setLocation,
+        setPromptDismissed,
+        setLocationExpired,
+        logAnalytics,
+        detectLocation,
+    });
 
-            return normalizeLocation(
-                {
-                    city: user.location.city,
-                    state: user.location.state,
-                    coordinates: user.location.coordinates,
-                    locationId: user.locationId || user.location.id,
-                    source: "manual",
-                    name: user.location.city,
-                    display: user.location.city,
-                },
-                "manual"
-            );
-        } catch {
-            return null;
-        }
-    }, [initialHasAuthCookie]);
-
-    const readPromptDismissedFromStorage = () => {
-        if (typeof window === "undefined") return false;
-        
-        const isDismissed = localStorage.getItem(LOCATION_PROMPT_DISMISSED_KEY) === "true";
-        if (!isDismissed) return false;
-        
-        // Check 30-day expiry for "Not Now"
-        const expiryStr = localStorage.getItem(LOCATION_PROMPT_DISMISSED_KEY + "_expiry");
-        if (expiryStr) {
-            const expiry = parseInt(expiryStr, 10);
-            if (!isNaN(expiry) && Date.now() > expiry) {
-                // Expired! Reset it.
-                localStorage.removeItem(LOCATION_PROMPT_DISMISSED_KEY);
-                localStorage.removeItem(LOCATION_PROMPT_DISMISSED_KEY + "_expiry");
-                return false;
-            }
-        }
-        
-        return true;
-    };
-
-    useEffect(() => {
-        if (initializedRef.current) return;
-        initializedRef.current = true;
-
-        let cancelled = false;
-
-        const initLocation = async () => {
-            setStatus("checking");
-            
-            const isDismissed = readPromptDismissedFromStorage();
-            setPromptDismissed(isDismissed);
-            
-            const permBlocked = readPermissionBlockedFlag();
-            if (permBlocked) {
-                setStatus("denied");
-            }
-
-            const hadStoredRaw = typeof window !== "undefined" && Boolean(localStorage.getItem("esparex_location"));
-            const storedLocation = readStoredLocation();
-            if (storedLocation) {
-                applyResolvedLocation(storedLocation, true);
-                return;
-            }
-            if (hadStoredRaw) {
-                setLocationExpired(true);
-            }
-
-            const profileLocation = await hydrateProfileLocation();
-            if (cancelled) return;
-
-            if (profileLocation) {
-                applyResolvedLocation(profileLocation, true);
-                return;
-            }
-
-            setLocation(DEFAULT_APP_LOCATION);
-            
-            // Capability-based permission check
-            let browserState = "prompt";
-            if (typeof navigator !== "undefined" && navigator.permissions) {
-                try {
-                    const result = await navigator.permissions.query({ name: 'geolocation' });
-                    browserState = result.state;
-                    
-                    // Listen for changes natively
-                    result.onchange = () => {
-                        if (result.state === "denied") {
-                            setStatus("denied");
-                            setPermissionBlockedFlag(true);
-                        } else if (result.state === "granted") {
-                            // If granted out of band, trigger auto-detect
-                            detectLocation(true, true);
-                        }
-                    };
-                } catch {
-                    // Fallback to prompt if API fails (e.g. older Safari)
-                }
-            }
-
-            if (permBlocked || browserState === "denied") {
-                setStatus("denied");
-                setPermissionBlockedFlag(true);
-            } else if (!isDismissed) {
-                // Introduce the delay before prompting
-                promptDelayTimeoutRef.current = setTimeout(() => {
-                    if (!cancelled) {
-                        setStatus(prev => {
-                            if (prev === "checking" || prev === "unknown") {
-                                // Defer analytics to avoid side-effects in reducer
-                                setTimeout(() => {
-                                    logAnalytics?.({
-                                        source: 'default',
-                                        city: 'Unknown',
-                                        state: 'Unknown',
-                                        reason: 'initial_prompt_shown',
-                                        eventType: 'location_prompt_shown'
-                                    });
-                                }, 0);
-                                return "prompt";
-                            }
-                            return prev;
-                        });
-                    }
-                }, LOCATION_PROMPT_DELAY_MS);
-            } else {
-                setStatus("prompt"); // It's dismissed, but the state is still prompt/available to be activated manually
-            }
-        };
-
-        void initLocation();
-
-        return () => {
-            cancelled = true;
-            if (promptDelayTimeoutRef.current) clearTimeout(promptDelayTimeoutRef.current);
-        };
-    }, [applyResolvedLocation, hydrateProfileLocation, readStoredLocation, readPermissionBlockedFlag, setPermissionBlockedFlag, logAnalytics, detectLocation]);
-
-    // ── Multi-Tab Synchronization ─────────────────────────────────────────────
-    useEffect(() => {
-        const handleStorageChange = (e: StorageEvent) => {
-            if (e.key === 'esparex_app_location' && e.newValue) {
-                try {
-                    const newLocation = JSON.parse(e.newValue) as LocationData;
-                    setLocation(newLocation);
-                    setStatus(getLocationStatus(newLocation.source));
-                } catch {
-                    // Ignore parse errors
-                }
-            } else if (e.key === LOCATION_PROMPT_DISMISSED_KEY) {
-                setPromptDismissed(e.newValue === "true");
-            }
-        };
-
-        window.addEventListener('storage', handleStorageChange);
-        return () => window.removeEventListener('storage', handleStorageChange);
-    }, []);
+    useMultiTabLocationSync({
+        setLocation,
+        setStatus,
+        setPromptDismissed,
+        getLocationStatus,
+    });
 
     // ── Context Values ────────────────────────────────────────────────────────
 
@@ -455,7 +307,6 @@ export function LocationProvider({
         }),
         [status, detectError, isDetecting, locationExpired, detectFeedback, _promptDismissed]
     );
-
 
     const actionsValue = useMemo(
         () => ({
@@ -503,3 +354,4 @@ export function useLocationStatus(): LocationStatusContextType {
     if (!ctx) throw new Error("useLocationStatus must be used within LocationProvider");
     return ctx;
 }
+
